[CHORE] 위도, 경도 파라미터 타입 -> Double로 변경 (#36)

+++ CLD/Data/Network/API/GymsAPI.swift
@@ -11,7 +11,7 @@ import Moya
 
 enum GymsAPI {
     case getGyms(keyword: String, limit: Int, skip: Int)
-    case getLocationGyms(x: Int, y: Int, keyword: String, limit: Int, skip: Int)
+    case getLocationGyms(x: Double, y: Double, keyword: String, limit: Int, skip: Int)
 }
 
 extension GymsAPI: BaseTargetType {
